phase apk changes

+++ RAMS/Mobile/RAMMS.MobileApps/PageModel/FormDAddLabourPageModel.cs
@@ -62,7 +62,7 @@ namespace RAMMS.MobileApps.PageModel
 
         public string SelectedUnit { get; set; }
 
-        public int? strQty { get; set; }
+        public decimal? strQty { get; set; }
 
         public string strRemarksValue { get; set; }
 
@@ -860,5 +860,7 @@ namespace RAMMS.MobileApps.PageModel
                 });
             }
         }
+
+
     }
 }
